eco-buddies/Eco-Bot.py
import autogen as ag
import openai
import whisper
import cv2  # OpenCV for computer vision
# Import other necessary libraries
from PIL import Image
from icecream import ic
from agents.autogen_agents import GeneralManagerAgent, Agent, DigitalTwinAgent
# Load personality script
with open("eco-buddies/eco-bot_personality.js", "r") as file:
    personality_script = file.read()

class EcoBot:
    def __init__(self):
        self.general_manager = GeneralManagerAgent()
        self.personality = self.load_personality()

    @staticmethod
    def load_personality():
        with open("eco-buddies/eco-bot_personality.js", "r") as file:
            return file.read()

    def handle_input(self, user_input):
        response = self.general_manager.manage_conversation(user_input)
        return response

# ...

def process_image(image_path):
    image = cv2.imread(image_path)
    gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    # ... other image processing steps
    return gray_image  # or other processed image data

# ...


# In a suitable module or part of eco-bot.py

def transcribe_audio(audio_path):
    # Assume whisper_transcribe is a function from the Whisper library
    transcription = whisper_transcribe(audio_path)
    return transcription

# ...

def manage_conversation(self, user_input):
    ecobot = EcoBot()
    # Use Eco-Bot to process user input and manage conversation
    # ...
